+++ NEWS.md
@@ -1,3 +1,5 @@
+# withr (development version)
+
 # withr 3.0.0
 
 ## Performance of withr