+++ NEWS.md
@@ -36,7 +36,7 @@
 
 * `deferred_run()` can now be run at any point in a knitr file (#235).
 
-,* `local_tempfile()` now writes `lines` in UTF-8 (#210) and always uses 
+* `local_tempfile()` now writes `lines` in UTF-8 (#210) and always uses 
   `\n` for newlines (#216).
 
 * `local_pdf()` and friends now correctly restore to the previously